refactor: relocate CG video input types
