+++ models/gen_cg_video/__init__.py
@@ -1,6 +1,6 @@
 """Cinematic / CG video generation model wrappers."""
 
 from .seedance_model import SeedanceModel
-from .types import VideoGenerationInput, VideoGenerationMode
+from .utils import VideoGenerationInput, VideoGenerationMode
 
 __all__ = ["SeedanceModel", "VideoGenerationInput", "VideoGenerationMode"]
